Time: 0 ms (100.00%), Space: 6.2 MB (22.55%) - LeetHub

+++ 458-poor-pigs/458-poor-pigs.cpp
@@ -1,6 +1,6 @@
 class Solution {
 public:
     int poorPigs(int buckets, int minutesToDie, int minutesToTest) {
-        return ceil((double)log(buckets) / (double)log(minutesToTest/minutesToDie + 1));
+        return ceil(log(buckets) / log(minutesToTest/minutesToDie + 1));
     }
 };
